add comments for backend,  Input validation for backend displayed to frontend

+++ user/views.py
@@ -19,15 +19,18 @@ from SyncMore import settings
 
 from .forms import DocumentForm
 from .models import Document, Email, Note, Phone, Supervisor, User
-from django.shortcuts import get_object_or_404
+from django.contrib import messages
 
 sys.path.append('..')
 
 
+# This function is used to login
 def login_view(request):
+    # Display the page if getting a GET request
     if request.method == 'GET':
         if request.session.get('username') and request.session.get('uid'):
             return HttpResponseRedirect('/user/index')
+        # Get the session
         c_username = request.COOKIES.get('username')
         c_uid = request.COOKIES.get('uid')
         if c_username and c_uid:
@@ -36,6 +39,7 @@ def login_view(request):
             return HttpResponseRedirect('/user/index')
 
         return render(request, 'user/login.html')
+    # Handle the POST request if a user tries to login
     elif request.method == 'POST':
         username = request.POST['username']
         password = request.POST['password']
@@ -46,14 +50,19 @@ def login_view(request):
             print('--login user error %s' % (e))
             note = 'username or password incorrect'
             dis = 'block'
+            messages.error(request, 'Username or password incorrect')
+
             return render(request, 'user/login.html', locals())
 
+        # Encrypt the password
         m = hashlib.md5()
         m.update(password.encode())
 
+        # If the password doesn't match
         if m.hexdigest() != user.password:
             note = 'username or password incorrect'
             dis = 'block'
+            messages.error(request, 'Username or password incorrect')
             return render(request, 'user/login.html', locals())
 
         request.session['username'] = username
@@ -61,6 +70,7 @@ def login_view(request):
 
         resp = HttpResponseRedirect('/user/index')
 
+        # Set up the session time
         if 'remember' in request.POST:
             resp.set_cookie('username', username, 3600 * 24 * 3)
             resp.set_cookie('uid', user.id, 3600 * 24 * 3)
@@ -68,38 +78,49 @@ def login_view(request):
         return resp
 
 
+# This function is used to register for a new account
 def reg_view(request):
+    # Display the page if getting a GET request
     if request.method == 'GET':
         return render(request, 'user/register.html')
+    # Handle the POST request if a user tries to sign up for a new account
     elif request.method == 'POST':
         username = request.POST['username']
         password_1 = request.POST['password_1']
         password_2 = request.POST['password_2']
         second_password = request.POST['second_password']
 
+        # Password restriction
         if len(password_1) < 6:
             note = 'the length of password is too short, at least 6 letters'
             dis = 'block'
+            messages.error(request, 'The length of password is too short, at least 6 letters')
             return render(request, 'user/register.html', locals())
 
         if not re.search("^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).*$", password_1):
             note = 'Failed, the password should include at least one digit, one uppercase and one lower case'
             dis = 'block'
+            messages.error(request,
+                           'Failed, the password should include at least one digit, one uppercase and one lower case')
             return render(request, 'user/register.html', locals())
 
+        # If the two password don't match
         if password_1 != password_2:
             note = 'the two passwords are not match'
             dis = 'block'
+            messages.error(request, 'The two passwords are not match')
             return render(request, 'user/register.html', locals())
 
         m = hashlib.md5()
         m.update(password_1.encode())
         password_m = m.hexdigest()
 
+        # If the username has been used
         old_users = User.objects.filter(Username=username)
         if old_users:
             note = 'username has been signed up'
             dis = 'block'
+            messages.error(request, 'Username has been signed up')
             return render(request, 'user/register.html', locals())
 
         try:
@@ -108,6 +129,7 @@ def reg_view(request):
             print('--create user error %s' % (e))
             note = 'username has been signed up'
             dis = 'block'
+            messages.error(request, 'Username has been signed up')
             return render(request, 'user/register.html', locals())
 
         request.session['username'] = username
@@ -116,13 +138,15 @@ def reg_view(request):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to logout
 def logout_view(request):
+    # Get the session and clear it
     if 'username' in request.session:
         del request.session['username']
     if 'uid' in request.session:
         del request.session['uid']
     resp = HttpResponseRedirect('../')
-
+    # Get the COOKIES and clear it
     if 'username' in request.COOKIES:
         resp.delete_cookie('username')
     if 'uid' in request.COOKIES:
@@ -130,6 +154,7 @@ def logout_view(request):
     return resp
 
 
+# This function is used to generate the authentication url to modify the object storage
 def generate_presigned_url(object_name):
     s3_client = boto3.client('s3',
                              endpoint_url=settings.AWS_S3_ENDPOINT_URL,
@@ -146,7 +171,10 @@ def generate_presigned_url(object_name):
         print(e)
         return None  # Handle exceptions according to your needs
 
+
+# This function is used to display the index view
 def index_view(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
@@ -165,10 +193,11 @@ def index_view(request):
                                             type=document_type) if document_type else Document.objects.filter(
             document_user_id=c_uid)
 
+        # Generate the data passed to frontend
         documents_with_urls = []
         for document in documents:
             presigned_url = generate_presigned_url(
-                document.document.name)  # Assuming 'document' field is the file field in your Document model
+                document.document.name)
             documents_with_urls.append({
                 'id': document.id,
                 'title': document.title,
@@ -189,8 +218,7 @@ def index_view(request):
             'title': document.title,
             'url': presigned_url
         })
-
-
+    # Generate the data passed to frontend
     context = {
         'user': user,
         'phones': phones,
@@ -204,36 +232,48 @@ def index_view(request):
     return render(request, 'user/index.html', context)
 
 
+# This function is used to add the phone
 def add_phone(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/add_phone.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         phone = request.POST.get('phone', "")
         Phone.objects.create(phone_user_id=c_uid, phone=phone)
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to add the email
 def add_email(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/add_email.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         email = request.POST.get('email', "")
         Email.objects.create(email_user_id=c_uid, email=email)
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to add the note
 def add_note(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/add_note.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         title = request.POST.get('title', "")
         content = request.POST.get('content', "")
@@ -241,14 +281,18 @@ def add_note(request):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to add the document
 def add_document(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
+    # Display the page if getting a GET request
     if request.method == "GET":
         form = DocumentForm()
         # create dropdown options list and pass into render?
         return render(request, 'user/add_document.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         title = request.POST.get('title', "")
         document = request.FILES.get('document', "")
@@ -257,17 +301,22 @@ def add_document(request):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to delete the phone
 def delete_phone(request, phone_id):
+    # Make changes to the database if getting a POST request
     if request.method == "POST":
         phone = Phone.objects.get(id=phone_id)
         phone.delete()
     return HttpResponseRedirect('/user/index')
 
 
+# This function is used to modify the phone
 def modify_phone(request, phone_id):
     phone = Phone.objects.get(id=phone_id)
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/modify_phone.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         phonee = request.POST.get('phone', "")
         phone.phone = phonee
@@ -275,17 +324,22 @@ def modify_phone(request, phone_id):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to delete the email
 def delete_email(request, email_id):
+    # Make changes to the database if getting a POST request
     if request.method == "POST":
         email = Email.objects.get(id=email_id)
         email.delete()
     return HttpResponseRedirect('/user/index')
 
 
+# This function is used to modify the email
 def modify_email(request, email_id):
     email = Email.objects.get(id=email_id)
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/modify_email.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         emaill = request.POST.get('email', "")
         email.email = emaill
@@ -293,17 +347,22 @@ def modify_email(request, email_id):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to delete the note
 def delete_note(request, note_id):
+    # Make changes to the database if getting a POST request
     if request.method == "POST":
         note = Note.objects.get(id=note_id)
         note.delete()
     return HttpResponseRedirect('/user/index')
 
 
+# This function is used to modify the note
 def modify_note(request, note_id):
     note = Note.objects.get(id=note_id)
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/modify_note.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         title = request.POST.get('title', "")
         content = request.POST.get('content', "")
@@ -313,6 +372,7 @@ def modify_note(request, note_id):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to delete the document from the object storage
 def delete_object_from_r2(object_name):
     delete_url = f"{settings.AWS_S3_ENDPOINT_URL}/{settings.AWS_STORAGE_BUCKET_NAME}/{object_name}"
     headers = {
@@ -323,20 +383,24 @@ def delete_object_from_r2(object_name):
     return response.status_code == 200
 
 
+# This function is used to delete the document
 def delete_document(request, document_id):
+    # Make changes to the database if getting a POST request
     if request.method == "POST":
         document = Document.objects.get(id=document_id)
         object_name = document.document.name
-        print(object_name)
-        delete_object_from_r2(object_name)
-        document.delete()
+        if delete_object_from_r2(object_name):
+            document.delete()
     return HttpResponseRedirect('/user/index')
 
 
+# This function is used to modify the document
 def modify_document(request, document_id):
     document = Document.objects.get(id=document_id)
+    # Display the page if getting a GET request
     if request.method == "GET":
         return render(request, 'user/modify_document.html', locals())
+    # Make changes to the database if getting a POST request
     elif request.method == "POST":
         title = request.POST.get('title', "")
         documentt = request.FILES.get('document', document.document)
@@ -346,7 +410,9 @@ def modify_document(request, document_id):
         return HttpResponseRedirect('/user/index')
 
 
+# This function is used to display the user account information
 def account(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
@@ -355,11 +421,14 @@ def account(request):
     return render(request, 'user/account.html', locals())
 
 
+# This function is used to modify the pin for the personal drive
 def modify_second_password(request):
+    # Get the session
     c_uid = request.COOKIES.get('uid')
     if c_uid is None:
         c_uid = request.session['uid']
     user = User.objects.get(id=c_uid)
+    # Save the new pin
     second_password = request.POST.get('second_password', "")
     user.second_password = second_password
     user.save()
